--- source/features/search/contexts.ts
@@ -1,4 +1,0 @@
-import React from 'react';
-import { ISearchContext } from './types';
-
-export const searchContext = React.createContext<ISearchContext | null>(null);
