--- source/features/search/types.ts
@@ -1,9 +1,0 @@
-import { SearchActions } from './actions';
-import { SearchApi } from './api';
-import { SearchStore } from './store';
-
-export interface ISearchContext {
-  actions: SearchActions;
-  api: SearchApi;
-  store: SearchStore;
-}
